refactor: TestRunsSearch is its own component and is visible on both the TestRunDetails (details page) and TestRunsDetails (table page)
