+++ galasa-ui/src/styles/test-runs/test-run-details/TestRun.module.css
@@ -5,7 +5,7 @@
  */
 .testRunContainer {
   width: 90%;
-  margin: 5rem auto 2rem auto;
+  margin: 2.5rem auto 2rem auto;
 }
 
 .tabs {
